+++ packages/eslint-plugin-fiori-tools/eslintrc-typescript.js
@@ -1,5 +1,5 @@
 'use strict';
-const { getResourcePaths } = require('../../lib/utils');
+const { getResourcePaths } = require('./lib/utils');
 const { sourceCodePath, testCodePath } = getResourcePaths();
 const overrides = [
     {
@@ -14,7 +14,6 @@ const overrides = [
             `${testCodePath}/**/Example.qunit.ts`,
             'backup/**',
             '*.d.ts',
-            
             '**/*.d.ts'
         ],
         'parser': '@typescript-eslint/parser',