+++ packages/eslint-plugin-fiori-tools/test/index.test.ts
@@ -36,7 +36,7 @@ describe('Tests for utils.ts', () => {
     });
 
     test('configs', async () => {
-        // eslint-disable-next-line @typescript-eslint/no-require-imports
+        // eslint-disable-next-line @typescript-eslint/no-var-requires
         const configs = require('../src/index');
         expect(configs).toBeDefined();
     });