+++ packages/preview-middleware-client/eslint.config.js
@@ -2,10 +2,12 @@ const { defineConfig } = require('eslint/config');
 const js = require('@eslint/js');
 
 const jsdoc = require('eslint-plugin-jsdoc');
-const config  = require('@sap-ux/eslint-plugin-fiori-tools');
+// TODO re-enable `@sap-ux/eslint-plugin-fiori-tools` after updating to eslint v9
+// const config  = require('@sap-ux/eslint-plugin-fiori-tools');
 
 module.exports = defineConfig([
-    ...config.defaultTS,
+// TODO re-enable `@sap-ux/eslint-plugin-fiori-tools` after updating to eslint v9
+    // ...config.defaultTS,
     {
         languageOptions: {
             ecmaVersion: 5,